myceliumd[-private]: add --log-format option
In addition to the previous format "compact", which is the default, this
can be set to "logfmt".

This isn't colorful yet, so `ansi_logs` is not enabled, as we'd only
want that for interactive usage, not when run via a service manager.

There's some amount of duplication between the two CLI entrypoints, but
I was a bit unsure where to put the shared code for tracing setup.
Happy to move this elsewhere too.